@@ -8,36 +8,36 @@
     "build": "rm -rf dist && NODE_ENV=production webpack --config webpack.build.js --display-modules"
   },
   "devDependencies": {
-    "autoprefixer-loader": "~2.1.0",
-    "babel-core": "~5.8.22",
-    "babel-eslint": "~4.0.10",
+    "autoprefixer-loader": "~3.1.0",
+    "babel-core": "~5.8.23",
+    "babel-eslint": "~4.1.1",
     "babel-loader": "~5.3.2",
+    "babel-runtime": "~5.8.20",
     "chai": "~3.2.0",
-    "css-loader": "~0.16.0",
-    "eslint": "~1.2.0",
+    "css-loader": "~0.17.0",
+    "eslint": "~1.3.1",
     "eslint-loader": "~1.0.0",
-    "eslint-plugin-angular": "~0.4.0",
     "extract-text-webpack-plugin": "~0.8.2",
     "file-loader": "~0.8.4",
-    "html-webpack-plugin": "~1.6.0",
+    "html-webpack-plugin": "~1.6.1",
     "karma": "~0.13.9",
-    "karma-coverage": "~0.5.0",
+    "karma-coverage": "~0.5.2",
     "karma-mocha": "~0.2.0",
     "karma-phantomjs-launcher": "~0.2.1",
     "karma-sinon-chai": "~1.0.0",
     "karma-webpack": "~1.7.0",
-    "mocha": "~2.2.5",
-    "ng-annotate-loader": "0.0.6",
-    "ng-cache-loader": "0.0.9",
+    "mocha": "~2.3.2",
+    "ng-annotate-loader": "0.0.10",
+    "ng-cache-loader": "0.0.13",
     "node-bourbon": "~4.2.3",
     "node-libs-browser": "~0.5.2",
-    "node-sass": "~3.3.0",
-    "phantomjs": "~1.9.17",
+    "node-sass": "~3.3.2",
+    "phantomjs": "~1.9.18",
     "raw-loader": "~0.5.1",
     "sass-loader": "~2.0.1",
     "style-loader": "~0.12.3",
     "url-loader": "~0.5.6",
-    "webpack": "~1.11.0",
+    "webpack": "~1.12.1",
     "webpack-dev-server": "~1.10.1",
     "webpack-notifier": "~1.2.1"
   },
@@ -55,5 +55,9 @@
   ],
   "author": "zackyang",
   "license": "MIT",
-  "homepage": "https://github.com/zack9433/angular-sanji-window"
+  "homepage": "https://github.com/zack9433/angular-sanji-window",
+  "dependencies": {
+    "angular": "~1.4.5",
+    "angular-material": "~0.11.0"
+  }
 }
