@@ -1,10 +1,13 @@
+import ngMaterial from 'angular-material';
+
 import './sanji-window.scss';
 import './sanji-window.tpl.html';
 import SanjiWindowService from './sanji-window.service';
 import SanjiWindowController from './sanji-window.controller';
 import SanjiWindowDirective from './sanji-window.directive';
 
-angular.module('sanji.window', ['ngMaterial'])
-  .service('sanjiWindowService', SanjiWindowService)
-  .controller('SanjiWindowController', SanjiWindowController)
-  .directive('sanjiWindow', SanjiWindowDirective.directiveFactory);
+let app = angular.module('sanji.window', [ngMaterial]);
+app.service('sanjiWindowService', SanjiWindowService);
+app.controller('SanjiWindowController', SanjiWindowController);
+app.directive('sanjiWindow', SanjiWindowDirective.directiveFactory);
+export default app = app.name
